Create basic project files. First iteration
First, a p5.js sketch is created with a setup() function that uses the createCanvas() function to create a canvas. The canvas size is set to 600x600 pixels in this example.

Next, a createGrid() function is defined to take two arguments: numRows and numCols, which indicate the number of rows and columns in the grid.

Inside the createGrid() function, the width and height of each cell in the grid are calculated based on the canvas size and the number of rows and columns.

An array of colors is created to use with the grid. A gradient of six colors is used in this example, but the number of colors can be adjusted.

A nested for loop is used to iterate over each cell in the grid. For each cell, its x and y position is calculated based on its row and column index. Then, Perlin noise is used to generate a random color value for the cell. The Perlin noise value is mapped to an index in the colors array to select a fill color for the cell.

Finally, the createGrid() function is called from the draw() function of the p5.js sketch. A Perlin noise animation is added by incrementing the noise value over time.

Optionally, CSS can be added to center the canvas in the browser window. This is achieved by setting the position to absolute and using the top, left, and transform properties to center the canvas within its parent container.